refactor(EasyTrash): Renamed module to InventoryCleaner.kt and add more descriptions to settings
